Moved build of generically correct data to separate file
