+++ PopulateDatabase.py
@@ -1,3 +1,5 @@
+"""Populate the database with a complete set of randomly generated data"""
+
 import os
 from random import choice, randint
 from typing import Any, List, Optional, Union
@@ -11,10 +13,8 @@ django.setup()
 import kiosk.CheckoutLogic as logic
 import names
 import progressbar
-from core.models.CertificationModels import Certification
 from core.models.DepartmentModels import Department
 from core.models.MemberModels import Member, Staffer
-from core.models.QuizModels import Question, Answer
 from core.models.TransactionModels import Transaction
 from core.models.GearModels import GearType, CustomDataField
 from django.contrib.auth.models import Group
@@ -22,7 +22,8 @@ from django.core.exceptions import ValidationError
 from django.db.utils import IntegrityError
 from django.utils.timezone import timedelta
 
-import buildPermissions
+import buildBasicData
+buildBasicData.build_all()
 
 ADMIN_RFID = '0000000000'
 SYSTEM_RFID = '1111111111'
@@ -89,29 +90,6 @@ def generate_rand_member() -> Member:
     return random_member
 
 
-def save_question(question_name=None, question_text=None, choices=None, correct_answer_index=0, error_message="You're Wrong!"):
-
-    # Make and save all of the answers
-    answers = []
-    for answer in choices:
-        ans = Answer(answer_phrase=answer[0], answer_text=answer[1])
-        ans.save()
-        answers.append(ans)
-
-    # make the question and relate all the answers
-    question = Question.objects.create(
-        usage="membership",
-        name=question_name,
-        question_text=question_text,
-        correct_answer=answers[correct_answer_index],
-        error_message=error_message
-    )
-    question.answers.add(*answers)
-    question.save()
-
-# Build all the groups and permissions
-buildPermissions.build_all()
-
 # Add the master admin  and excursion system accounts
 admin = Member.objects.create_superuser(
     "admin@excursionclubucsb.org", ADMIN_RFID, password=PASSWORD
@@ -175,42 +153,6 @@ staffer.save()
 print('')
 print('Made staffers')
 
-# Add certifications
-kayak_cert = Certification(
-    title='Kayaking', requirements='1) Be able to swim god dammit '
-    '2) Have received the safety rant, know about wind and current '
-    '3) Be able to take the kayak out safely '
-    '4) Be able to get off of, flip, and get into a kayak out in the water '
-    '5) Be able to bring the kayak back in to shore safely'
-)
-kayak_cert.save()
-
-sup_cert = Certification(
-    title='Stand Up Paddleboarding', requirements='1) Be able to swim god dammit '
-    '2) Have received the safety rant, know about wind and current '
-    '3) Be able to take the SUP out safely '
-    '4) Be able to get off of, flip, and get into a SUP out in the water '
-    '5) Be able to bring the SUP back in to shore safely'
-)
-sup_cert.save()
-
-# Add departments
-departments = [
-    'Camping', 'Backpacking', 'Rock Climbing', 'Skiing/Snowboarding', 'Kayaking', 'Paddleboarding',
-    'Surfing', 'Wetsuits', 'Mountaineering', 'Archery', 'Paintballing', 'Free Diving', 'Off-Road'
-]
-all_staffers = Staffer.objects.all()
-for dept in departments:
-    name = dept
-    details = 'All the gear related to {}'.format(name)
-    stl: str = pick_random(all_staffers)
-    department = Department(name=name, description=details)
-    department.save()
-    department.stls.add(stl)
-    department.save()
-print('')
-print('Made departments')
-
 # Add custom fields and gear_types
 field_data = {
     "length": {
@@ -241,7 +183,7 @@ field_data = {
         "required": True,
         "help_text": "The size of the item",
         "choices": (
-            ("--", "Please choose a size"),
+            ("", "Please choose a size"),
             ("S", "Small"),
             ("M", "Medium"),
             ("L", "Large")
@@ -285,6 +227,8 @@ for field_name in field_data.keys():
     )
     if 'suffix' in field_data[field_name].keys():
         field.suffix = field_data[field_name]['suffix']
+    if 'choices' in field_data[field_name].keys():
+        field.choices = field_data[field_name]['choices']
     field.save()
     custom_fields.append(field)
 
@@ -377,69 +321,6 @@ for i in range(5):
     except ValidationError as e:
         pass
 
-print('Writing questions...')
 
-# Write quiz questions
-save_question(
-    question_name="Punishment",
-    question_text="What is the punishment for breaking a club rule?",
-    choices=(
-        ("wallow", "Wallow in you own incompetence"),
-        ("arrest", "We have you arrested"),
-        ("membership", "Forfeit your membership"),
-        ("lashes", "10 lashes before the mast")
-    ),
-    correct_answer_index=2,
-    error_message="If you break a rule you lose your membership!"
-)
-save_question(
-    question_name="gear",
-    question_text="How many of each item type can you check out?",
-    choices=(
-        ("None", "Trick question, we don't check out gear"),
-        ("1", "One of each type of gear"),
-        ("2", "Two of each type of gear"),
-        ("unlimited", "as many as you'd like")
-
-    ),
-    correct_answer_index=1,
-    error_message="You can only check out one of each type of item!"
-)
-save_question(
-    question_name="certification",
-    question_text="How do you get certified for kayaks and SUPS?",
-    choices=(
-        ("class", "Take a $500 class"),
-        ("trip", "Go on a trip with a staffer"),
-        ("date", "Bang a bunch of staffers"),
-        ("nudie", "Run naked around the block")
-    ),
-    correct_answer_index=1,
-    error_message="To get certified just go on a trip with a staffer."
-)
-save_question(
-    question_name="broken",
-    question_text="What do you do when a piece of gear breaks?",
-    choices=(
-        ("hide", "Hide it and hope no one notices"),
-        ("run", "Run away, Simba! Run away and NEVER return!"),
-        ("fine", "Pay a $10 fine for broken gear"),
-        ("tell", "Shit happens. Just let us know so we can fix it")
-    ),
-    correct_answer_index=3,
-    error_message="Just tell us it's broken so we can make sure it stays in good shape."
-)
-save_question(
-    question_name="staffers",
-    question_text="Who are our staffers?",
-    choices=(
-        ("volunteers", "Student volunteers who do this for fun"),
-        ("pros", "Well-paid professionals"),
-        ("blokes", "Random blokes we found on the street"),
-        ("fake", "Fake news. There are no staffers")
-    ),
-    correct_answer_index=0,
-    error_message="All our staffers are volunteers!"
-)
 
 print('Finished')
